functions/restingest/requirements.txt
azure-common==1.1.24
azure-functions==1.0.4
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
clickclick==1.2.2
connexion==2.0.2
cryptography==2.8
defusedxml==0.6.0
Flask==1.1.1
Flask-Cors==3.0.7
Flask-Testing==0.6.1
google-api-core==1.16.0
google-auth==1.11.0
google-cloud-core==1.3.0
google-cloud-kms==1.0.0
google-cloud-secret-manager==1.0.0
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.11.4
grpcio==1.26.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
jwkaas==0.4
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.1
oauthlib==3.1.0
openapi-spec-validator==0.2.8
pandas==0.24.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.6.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.14.0
swagger-ui-bundle==0.0.2
urllib3==1.25.8
Werkzeug==0.16.0
